Update area.cpp
El programa te fallaba y no te daba el área y el perímetro correctamente, habías puesto double, pero en la entrada de datos ponías %i, con lo cuál era imposible que funcionase, por eso te lo he cambiado y he puesto int.

Si quisieras usarlo con el double, te aconsejo que pongas cosas como "%d"

Aunque es más útil usar float, y poner "%f", ya que puedes poner eso o "%.2f", para poner dos decimales solo

+++ 01_ejemplos/area.cpp
@@ -1,10 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-double main(){
-    double lado,
-        area,
-        perimetro;
+int main(){
+   
+    int lado, area, perimetro;
 
     system("toilet --gay -fpagga CUADRADO");
